implements sending disconnect message on websocket failure
Adds an option _WS_RECONNECTS_BEFORE_ALERT after how many failed reconnects a disconnect
message will be send to the mqtt-topic PUBmcz/Status.
On connection of the websocket a message PUBmcz/Status with value connected is published
to the mqtt-broker.
If _WS_RECONNECTS_BEFORE_ALERT retries to reconnect to the websocket failed, a message
PUBmcz/disconnected is published to the mqtt-broker.

Changed case of global variables to lower-case following the naming conventions from
https://www.python.org/dev/peps/pep-0008/#global-variable-names.