improves running as system service
Systemd services shouldn't be run as root user.
This commit changes the install_daemon and update_daemon to create
a dedicated maestro_service user and group.
The location of the service is changed from the /opt/maestro directory
to the better suitable /usr/local/lib/maestro_service directory.
The maestro.py file adds handling of logging to journald and disables
logging to file when run as service.
The needed dependencies for system, and journal logging are added to
the install and update  scripts.